+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Fixed password fields not being detected as changed ([#7670](https://github.com/pocketbase/pocketbase/issues/7670)).
 
+- Added the local time zone name next to the `date` field label.
+
 
 ## v0.37.4